💫 UPDATE: migrate ListEmptyContent to function component with TS support
