--- client/reader/list-stream/empty.jsx
@@ -1,46 +1,0 @@
-import { localize } from 'i18n-calypso';
-import { Component } from 'react';
-import { connect } from 'react-redux';
-import EmptyContent from 'calypso/components/empty-content';
-import { recordAction, recordGaEvent } from 'calypso/reader/stats';
-import { recordReaderTracksEvent } from 'calypso/state/reader/analytics/actions';
-
-class ListEmptyContent extends Component {
-	shouldComponentUpdate() {
-		return false;
-	}
-
-	recordAction = () => {
-		recordAction( 'clicked_following_on_empty' );
-		recordGaEvent( 'Clicked Following on EmptyContent' );
-		this.props.recordReaderTracksEvent( 'calypso_reader_following_on_empty_list_stream_clicked' );
-	};
-
-	render() {
-		/* eslint-disable wpcalypso/jsx-classname-namespace */
-		const action = (
-			<a
-				className="empty-content__action button is-primary"
-				onClick={ this.recordAction }
-				href="/read"
-			>
-				{ this.props.translate( 'Back to Following' ) }
-			</a>
-		);
-
-		return (
-			<EmptyContent
-				title={ this.props.translate( 'No recent posts' ) }
-				line={ this.props.translate( 'The sites in this list have not posted anything recently.' ) }
-				action={ action }
-				illustration="/calypso/images/illustrations/illustration-empty-results.svg"
-				illustrationWidth={ 400 }
-			/>
-		);
-		/* eslint-enable wpcalypso/jsx-classname-namespace */
-	}
-}
-
-export default connect( null, {
-	recordReaderTracksEvent,
-} )( localize( ListEmptyContent ) );
